Relax Jakarta / Javax rules
The rules now know which Jakarta version made the package
change from 'javax' to 'jakarta'. Only the versions that
have the same package as the corresponding Javax components
are now in conflict.

The rules setup follows the pattern for all rules. There is one
rule per capability. So there are now sometimes two rules to detect
conflicts between APIs and different implementations of a certain
Javax/Jakarta component:
- Javax...Rule (for javax capability)
- Jakarta...Rule (for jakrata capability)

Related to #6