Support all future Guava versions
As the dependencies of the 'jre' and 'android' variants are now the
same, we do not need to adjust them anymore and hence to not need
to encode versions in the plugin.
Lets assume that it stays like this for future releases and also
assume that https://github.com/google/guava/pull/3683 will ever be
integrated.